refactor(optimize): split optimize_unified.rs into likelihood, zero_boundary, and dispatch
- likelihood.rs: OptimizationMetrics, evaluate_mixed, evaluate_with_indels
- zero_boundary.rs: is_zero_branch_optimal, reconcile_zero_boundary, grid search
- dispatch.rs: run_optimize_mixed, initial_guess_mixed, per-edge dispatch
- Newton tolerance functions moved to method_newton.rs alongside the constants they use
