+++ packages/treetime/src/optimize/__tests__/test_convergence/test_convergence_edge_cases.rs
@@ -1,7 +1,7 @@
 #[cfg(test)]
 mod tests {
   use crate::optimize::params::BranchOptMethod;
-  use crate::optimize::optimize_unified::run_optimize_mixed;
+  use crate::optimize::dispatch::run_optimize_mixed;
   use crate::representation::payload::ancestral::GraphAncestral;
   use eyre::Report;
   use rstest::rstest;
