+++ packages/treetime/src/optimize/__tests__/test_convergence_sc2.rs
@@ -5,7 +5,7 @@ mod tests {
   use crate::ancestral::marginal::update_marginal;
   use crate::gtr::get_gtr::{JC69Params, jc69};
   use crate::optimize::params::BranchOptMethod;
-  use crate::optimize::optimize_unified::initial_guess_mixed;
+  use crate::optimize::dispatch::initial_guess_mixed;
   use crate::optimize::run_loop::{collect_optimize_partitions, run_optimize_loop};
   use crate::representation::partition::fitch::PartitionFitch;
