+++ packages/treetime/src/optimize/__tests__/test_grid_search/test_grid_search_coverage.rs
@@ -1,6 +1,6 @@
 #[cfg(test)]
 mod tests {
-  use crate::optimize::optimize_unified::grid_search_branch_lengths;
+  use crate::optimize::zero_boundary::grid_search_branch_lengths;
   use approx::assert_abs_diff_eq;
   use ndarray::Array2;
   use rstest::rstest;
