+++ packages/treetime/src/optimize/__tests__/test_initial_guess_indel_zero_bl.rs
@@ -3,7 +3,7 @@ mod tests {
   use crate::alphabet::alphabet::{Alphabet, AlphabetName};
   use crate::ancestral::marginal::{initialize_marginal, update_marginal};
   use crate::gtr::get_gtr::{JC69Params, jc69};
-  use crate::optimize::optimize_unified::initial_guess_mixed;
+  use crate::optimize::dispatch::initial_guess_mixed;
   use crate::representation::partition::marginal_dense::PartitionMarginalDense;
   use crate::representation::payload::ancestral::GraphAncestral;
   use crate::seq::alignment::get_common_length;
