+++ packages/treetime/src/optimize/__tests__/test_newton_convergence/test_newton_convergence_tolerance.rs
@@ -1,9 +1,8 @@
 #[cfg(test)]
 mod tests {
   use crate::optimize::method_newton::NEWTON_REL_TOL;
-  use crate::optimize::optimize_unified::{
-    evaluate_mixed, newton_tolerance_log, newton_tolerance_sqrt, newton_tolerance_t,
-  };
+  use crate::optimize::likelihood::evaluate_mixed;
+  use crate::optimize::method_newton::{newton_tolerance_log, newton_tolerance_sqrt, newton_tolerance_t};
   use approx::assert_abs_diff_eq;
   use ndarray::array;
   use num::clamp;
