+++ packages/treetime/src/optimize/__tests__/test_no_indels.rs
@@ -8,7 +8,7 @@ mod tests {
   };
   use crate::optimize::params::{BranchOptMethod, InitialGuessMode};
   use crate::optimize::iteration::apply_initial_guess_mode;
-  use crate::optimize::optimize_unified::run_optimize_mixed_inner;
+  use crate::optimize::dispatch::run_optimize_mixed_inner;
   use crate::optimize::run_loop::run_optimize_loop;
   use crate::representation::partition::marginal_dense::PartitionMarginalDense;
   use crate::representation::payload::ancestral::GraphAncestral;
