+++ packages/treetime/src/optimize/__tests__/test_optimize_zero_sequence_length.rs
@@ -3,9 +3,7 @@ mod tests {
   use crate::alphabet::alphabet::{Alphabet, AlphabetName};
   use crate::gtr::get_gtr::{JC69Params, jc69};
   use crate::optimize::params::BranchOptMethod;
-  use crate::optimize::optimize_unified::{
-    initial_guess_mixed, run_optimize_mixed, run_optimize_mixed_with_indel_rate,
-  };
+  use crate::optimize::dispatch::{initial_guess_mixed, run_optimize_mixed, run_optimize_mixed_with_indel_rate};
   use crate::optimize::run_loop::collect_optimize_partitions;
   use crate::representation::partition::marginal_dense::PartitionMarginalDense;
   use crate::representation::partition::marginal_sparse::PartitionMarginalSparse;
