+++ packages/treetime/src/optimize/mod.rs
@@ -2,12 +2,14 @@
 mod __tests__;
 
 pub mod dense_eval;
+pub mod dispatch;
 pub mod eval;
 pub mod indel;
 pub mod iteration;
+pub mod likelihood;
 pub mod method_brent;
 pub mod method_newton;
-pub mod optimize_unified;
 pub mod params;
 pub mod run_loop;
 pub mod sparse_eval;
+pub mod zero_boundary;
