+++ packages/treetime/src/optimize/run_loop.rs
@@ -4,7 +4,7 @@ use crate::optimize::iteration::{
   apply_damping, find_zero_optimal_internal_edges, prune_and_merge_in_loop, restore_branch_lengths, save_branch_lengths,
 };
 use crate::optimize::indel::{estimate_indel_rate, total_indel_log_lh};
-use crate::optimize::optimize_unified::run_optimize_mixed_inner;
+use crate::optimize::dispatch::run_optimize_mixed_inner;
 use crate::representation::partition::marginal_dense::PartitionMarginalDense;
 use crate::representation::partition::marginal_sparse::PartitionMarginalSparse;
 use crate::representation::partition::traits::{PartitionOptimizeOps, PartitionOptimizeVec};
