+++ packages/treetime/src/timetree/inference/branch_length_likelihood.rs
@@ -1,5 +1,5 @@
 use crate::make_error;
-use crate::optimize::optimize_unified::evaluate_with_indels_log_lh_only;
+use crate::optimize::likelihood::evaluate_with_indels_log_lh_only;
 use crate::representation::partition::optimization_contribution::OptimizationContribution;
 use eyre::Report;
 use ndarray::Array1;
